requirements.txt
# Dépendances pour comprehension_orale

# API et services
openai>=1.0.0
python-dotenv>=0.19.0

# Text-to-Speech
azure-cognitiveservices-speech>=1.31.0
gtts>=2.2.4
edge-tts>=6.1.0

# Utilities
requests>=2.28.0
